pyproject.toml
[project]
name = "pythonproject3"
version = "0.1.0"
description = ""
authors = [{name = "Evgenii", email ="evgeniicebanica94@gmail.com"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.1"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^7.0.0"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = ".git"

[tool.isort]
# максимальная длина строки
line_length = 119




